+++ CHANGELOG.md
@@ -10,6 +10,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - Added helper functions for checking the error type of a write query, this is useful when calling `RunWrite`.
     + Added `IsConflictErr` which returns true when RethinkDB returns a duplicate key error.
     + Added `IsTypeErr` which returns true when RethinkDB returns an unexpected type error.
+- Added the `RawQuery` function which can be used to execute a raw JSON query, for more information about this query see the godoc.
 
 ## v1.3.2 - 2015-02-01